document-summarizer/requirements.txt
django
djangorestframework
pypdf2
django-cors-headers
reportlab